pyproject.toml
[tool.poetry]
name = "SentenX"
version = "0.1.0"
description = ""
authors = ["Rafael Soares <rafael.soares@weni.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
langchain = "^0.0.281"
tiktoken = "^0.4.0"
elasticsearch = "^8.9.0"
boto3 = "^1.28.41"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
huggingface-hub = "^0.16.4"
pypdf = "^3.17.0"
python-docx = "^1.1.0"
unstructured = "^0.10.29"
docx2txt = "^0.8"
networkx = "^3.2.1"
pandas = "^2.1.2"
openpyxl = "^3.1.2"


[tool.poetry.group.dev.dependencies]
coverage = "^7.3.1"
black = "^23.9.1"
reportlab = "^4.0.7"
xlsxwriter = "^3.1.9"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
